+++ R/calculateExpectedRT.R
@@ -1,11 +1,11 @@
 #############################################################################
-#' Calculate Expected Response Times
+#' Calculate Expected Response Times Moments
 #'
-#' Calculate expected response times given item parameters of the log normal response time model.
+#' Calculate expected response times moments (mean and variance) given item parameters of the log normal response time model.
 #'
 #' Expected response times are calculated according to the log normal response time model by
 #' van der Linden (2006) or Klein Entink et al. (2009). If \code{phi} is \code{1}, the model
-#' by van der Linden (2006) is used. Either a single set of parameters of vectors of each parameters
+#' by van der Linden (2006) is used. Either a single set of parameters of vectors of each parameter
 #' can be supplied.
 #'
 #' The calculation is based on Fenton (1960). For the model by van der Linden (2006), the calculation was
@@ -35,15 +35,21 @@
 #'@examples
 #'# expected RT for a single item (van der Linden model)
 #'calculateExpectedRT(lambda = 3.8, zeta = 0, sdEpsi = 0.3)
+#'calculateExpectedRTvar(lambda = 3.8, zeta = 0, sdEpsi = 0.3)
 #'
 #'# expected RT for multiple items (van der Linden model)
 #'calculateExpectedRT(lambda = c(4.1, 3.8, 3.5), zeta = 0,
 #'                    sdEpsi = c(0.3, 0.4, 0.2))
+#'calculateExpectedRTvar(lambda = c(4.1, 3.8, 3.5), zeta = 0,
+#'                    sdEpsi = c(0.3, 0.4, 0.2))
 #'
-#'# TIF for multiple items and multiple ability levels (1PL model)
+#'# expected RT for multiple items and multiple spped levels (Klein Entink model)
 #'calculateExpectedRT(lambda = c(3.7, 4.1, 3.8), phi = c(1.1, 0.8, 0.5),
 #'                     zeta = c(-1, 0, 1), sdEpsi = c(0.3, 0.4, 0.2))
+#'calculateExpectedRTvar(lambda = c(3.7, 4.1, 3.8), phi = c(1.1, 0.8, 0.5),
+#'                     zeta = c(-1, 0, 1), sdEpsi = c(0.3, 0.4, 0.2))
 #'
+#' @describeIn calculateExpectedRT Calculate expected mean response time
 #'@export
 calculateExpectedRT <- function(lambda, phi = rep(1, length(lambda)), zeta, sdEpsi) {
   if(!is.numeric(lambda)) stop("'lambda' must be a numeric vector.")
@@ -54,6 +60,19 @@ calculateExpectedRT <- function(lambda, phi = rep(1, length(lambda)), zeta, sdEp
 
   cumulants <- getCumulantRT(zeta, lambda, phi, sdEpsi)
   return(cumulants[[1]])
+}
+
+#' @describeIn calculateExpectedRT Calculate expected response time variance
+#'@export
+calculateExpectedRTvar <- function(lambda, phi = rep(1, length(lambda)), zeta, sdEpsi) {
+  if(!is.numeric(lambda)) stop("'lambda' must be a numeric vector.")
+  if(!is.numeric(phi)) stop("'phi' must be a numeric vector.")
+  if(!is.numeric(zeta)) stop("'zeta' must be a numeric vector.")
+  if(!is.numeric(sdEpsi)) stop("'sdEpsi' must be a numeric vector.")
+  if(length(sdEpsi) != length(lambda) || length(sdEpsi) != length(phi)) stop("'lambda', 'phi', and 'sdEpsi' must be of the same length.")
+
+  cumulants <- getCumulantRT(zeta, lambda, phi, sdEpsi)
+  return(cumulants[[2]])
 }